clean code rename Query claas fields
