+++ src/main/java/Main.java
@@ -1,6 +1,6 @@
 import search_engine.Document;
 import search_engine.SearchEngine;
-import search_engine.queryDecoders.CommonQueryDecoder;
+import search_engine.queryDecoder.CommonQueryDecoder;
 import search_engine.tokenizers.SpaceTokenizer;
 
 import java.io.File;
