+++ src/main/java/search_engine/SearchEngine.java
@@ -2,9 +2,9 @@ package search_engine;
 
 import com.google.common.collect.ImmutableSet;
 import search_engine.filters.Filter;
-import search_engine.queryDecoders.CommonQueryDecoder;
-import search_engine.queryDecoders.Query;
-import search_engine.queryDecoders.QueryDecoder;
+import search_engine.queryDecoder.CommonQueryDecoder;
+import search_engine.queryDecoder.Query;
+import search_engine.queryDecoder.QueryDecoder;
 import search_engine.tokenizers.SpaceTokenizer;
 import search_engine.tokenizers.Tokenizer;
 
@@ -56,34 +56,34 @@ public class SearchEngine {
     public Set<String> getQueryResult(Query query) {
         Set<String> results = new HashSet<>();
 
-        if (query.compulsories().isEmpty()) {
+        if (query.includes().isEmpty()) {
             if (query.optionals().isEmpty()) {
-                if (!query.forbidden().isEmpty()) {
+                if (!query.excludes().isEmpty()) {
                     results = docs.stream().map(Document::getId).collect(Collectors.toSet());
                 }
             } else {
                 results = itemsUnion(query.optionals());
             }
         } else {
-            results = intersectionCompulsories(query.compulsories());
+            results = intersectionCompulsories(query.includes());
             if (!query.optionals().isEmpty()) {
                 Set<String> optionalIds = itemsUnion(query.optionals());
                 results.removeIf(s -> !optionalIds.contains(s)); // results &= optionalIds
             }
         }
 
-        Set<String> forbiddenIds = itemsUnion(query.forbidden());
-        return removeForbidden(results, forbiddenIds);
+        Set<String> excludesIds = itemsUnion(query.excludes());
+        return removeExcludes(results, excludesIds);
     }
 
-    private Set<String> removeForbidden(Set<String> base, Set<String> forbidden) {
+    private Set<String> removeExcludes(Set<String> base, Set<String> excludes) {
         Set<String> result = new HashSet<>();
-        base.stream().parallel().filter(id -> !forbidden.contains(id)).forEach(result::add);
+        base.stream().parallel().filter(id -> !excludes.contains(id)).forEach(result::add);
         return new HashSet<>(result);
     }
 
-    private Set<String> intersectionCompulsories(List<String> compulsories) {
-        Optional<String> baseWordOptional = findBaseWord(compulsories);
+    private Set<String> intersectionCompulsories(List<String> includes) {
+        Optional<String> baseWordOptional = findBaseWord(includes);
 
         if (baseWordOptional.isEmpty()) {
             return new HashSet<>();
@@ -93,7 +93,7 @@ public class SearchEngine {
 
         Set<String> result = new HashSet<>(invertedIndex.get(baseWord));
 
-        for (String compulsory : compulsories) {
+        for (String compulsory : includes) {
             Set<String> foundIds = invertedIndex.get(compulsory);
             result.removeIf(s -> !foundIds.contains(s)); // result &= foundIds
             if (result.isEmpty()) {
@@ -104,11 +104,11 @@ public class SearchEngine {
         return result;
     }
 
-    private Optional<String> findBaseWord(List<String> compulsories) {
+    private Optional<String> findBaseWord(List<String> includes) {
         int minSize = Integer.MAX_VALUE;
         String baseWord = "";
 
-        for (String compulsory : compulsories) {
+        for (String compulsory : includes) {
             Set<String> foundIds = invertedIndex.get(compulsory);
 
             if (foundIds == null || foundIds.isEmpty()) {
