+++ src/main/java/search_engine/queryDecoder/CommonQueryDecoder.java
@@ -1,4 +1,4 @@
-package search_engine.queryDecoders;
+package search_engine.queryDecoder;
 
 import com.google.common.collect.ImmutableList;
 
@@ -10,9 +10,9 @@ public class CommonQueryDecoder implements QueryDecoder {
     public Query decode(String query) {
         String[] strings = query.split("\\s+");
 
-        List<String> compulsories = new ArrayList<>();
+        List<String>  includes = new ArrayList<>();
         List<String> optionals = new ArrayList<>();
-        List<String> forbidden = new ArrayList<>();
+        List<String>  excludes = new ArrayList<>();
 
         for (String string : strings) {
             if (string.isEmpty()) {
@@ -20,11 +20,11 @@ public class CommonQueryDecoder implements QueryDecoder {
             }
             switch (string.charAt(0)) {
                 case '+' -> optionals.add(string.substring(1));
-                case '-' -> forbidden.add(string.substring(1));
-                default -> compulsories.add(string);
+                case '-' ->  excludes.add(string.substring(1));
+                default ->  includes.add(string);
             }
         }
 
-        return new Query(ImmutableList.copyOf(compulsories), ImmutableList.copyOf(optionals), ImmutableList.copyOf(forbidden));
+        return new Query(ImmutableList.copyOf( includes), ImmutableList.copyOf(optionals), ImmutableList.copyOf( excludes));
     }
 }
