+++ src/main/java/search_engine/queryDecoder/Query.java
@@ -1,0 +1,7 @@
+package search_engine.queryDecoder;
+
+import java.util.List;
+
+public record Query(List<String> includes, List<String> optionals,
+                    List<String> excludes) {
+}
