+++ src/main/java/search_engine/queryDecoder/QueryDecoder.java
@@ -1,4 +1,4 @@
-package search_engine.queryDecoders;
+package search_engine.queryDecoder;
 
 public interface QueryDecoder {
     Query decode(String query);
