--- src/main/java/search_engine/queryDecoders/Query.java
@@ -1,9 +1,0 @@
-package search_engine.queryDecoders;
-
-import com.google.common.collect.ImmutableList;
-
-import java.util.List;
-
-public record Query(List<String> compulsories, List<String> optionals,
-                    List<String> forbidden) {
-}
